% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coverTreeNeighbor}
\alias{coverTreeNeighbor}
\title{Run a Cover Tree distance analysis}
\usage{
coverTreeNeighbor(dataset, k)
}
\arguments{
\item{dataset}{A matrix of training data values}

\item{k}{An integer specifying the number of classes}
}
\value{
A list with two elements giving the nighbors and their distances
}
\description{
Run a Cover Tree distance measure analysis which provides a specialised
}
\details{
This function performs a distance calculation using a Cover Tree.
A cover tree is a tree specifically designed to speed up nearest-neighbor
computation in high-dimensional spaces.  Each non-leaf node references a
point and has a nonzero number of children, including a \dQuote{self-child} which
references the same point.  A leaf node represents only one point.
}
