% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lars_dependent_x}
\alias{lars_dependent_x}
\title{LARS test data from the mlpack unit tests: regressors}
\format{A \sQuote{csv} file.}
\source{
\url{https://github.com/mlpack/mlpack/blob/master/src/mlpack/tests/data/lars_dependent_x.csv}
}
\description{
LARS test data from the mlpack unit tests: regressors
}
