% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{naiveBayesClassifier}
\alias{naiveBayesClassifier}
\title{Run a Naive Bayes Classifier}
\usage{
naiveBayesClassifier(train, labels, classes, test = NULL)
}
\arguments{
\item{train}{A matrix of training data values}

\item{labels}{An integer vector of target (class) labels, with the same 
length and the training data set}

\item{classes}{An integer with the given number of classes}

\item{test}{An optional test set, with the same number of columns as
the test set.}
}
\value{
A list with several components: means of training data means,
 variance of training data variances, probanbilities of training data
 class probabilities. If a test data set was supplied an estimated
 classification vector is also returned.
}
\description{
Run a Naive Bayes Classifier given training data (and optional test data).
}
\details{
This function uses train set matrix and vector of target labels to run
a naive Bayes classifier for an externally given nummber of class. If
an optional test data set is supplied, it will be evulated given the
estimated from the initial test set.
}
\examples{
data(trainSet)
M <- t(trainSet[, -5])    ## train data, transpose and removing class labels
lb <- trainSet[, 5]       ## class labels for train set
naiveBayesClassifier(M, lb, 2L)
tM <- t(testSet[, -5])    ## test data
naiveBayesClassifier(M, lb, 2L, tM)
}
