% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testSet}
\alias{testSet}
\title{Test data from the mlpack unit tests}
\format{A \sQuote{csv} file read in and \emph{transposed} by
 \code{mlpack} conventions.}
\source{
\url{https://raw.githubusercontent.com/mlpack/mlpack/master/src/mlpack/tests/data/testSet.csv}
}
\description{
Test data from the mlpack unit tests
}
