% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainRes}
\alias{trainRes}
\title{Training data result from the mlpack unit tests}
\format{A \sQuote{csv} file read in and \emph{transposed} by
 \code{mlpack} conventions.}
\source{
\url{https://raw.githubusercontent.com/mlpack/mlpack/master/src/mlpack/tests/data/trainRes.csv}
}
\description{
Training data result from the mlpack unit tests
}
