% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LARS}
\alias{LARS}
\title{Run a lars / lasso / elasticNet regression}
\usage{
LARS(matX, matY, testPoints, lambda1, lambda2, useCholesky = FALSE)
}
\arguments{
\item{data}{A matrix of data values}
}
\value{
A list with results
}
\description{
Run a lars/lasso/elasticNet regression
}
\details{
This function performs a lars, lasso or elastic net regression
}
\examples{
  # tbd
}
