% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LARS}
\alias{LARS}
\title{Run a lars / lasso / elasticNet regression}
\usage{
LARS(matX, vecY, lambda1, lambda2, useCholesky = FALSE, testX = NULL)
}
\arguments{
\item{matX}{A matrix of explanatory variables, i.e. regressors}

\item{vecY}{A vector with the target variable}

\item{lambda1}{A doube with l1-norm penalty regularization parameter}

\item{lambda2}{A doube with l2-norm penalty regularization parameter}

\item{useCholesky}{A logical value indicating whether to use the Cholesky}

\item{testX}{A optional matrix of test values to validate prediction
decomposition when solving the linear system, else full Gram matrix is used.}
}
\value{
A list with estimated coefficient, the value of lambda1 after each iteration
and the predicted values, either from the training data or, if supplied, the test
set.
}
\description{
Run a lars/lasso/elasticNet regression
}
\details{
This function performs a lars, lasso or elastic net regression.
}
\examples{
\dontrun{
  data(trainSet)
  mat <- trainSet[, -5]
  y <- trainSet[,5]
  tst <- testSet[, -5]
  res <- LARS(mat, y, tst, 0.5, 0.5)
}
}
